# 资料/backend-openapi.yaml
openapi: 3.1.0
info:
  title: Counseling Agent Backend API
  version: 1.2.0-sprint2-mood
  description: |
    心理咨询 Agent 后端接口文档。

    本文档覆盖 Sprint 1 闭环接口，以及 Sprint 2 知识库与情绪趋势 MVP：
    鉴权、文本对话、记忆管理、情绪记录、情绪趋势、安全资源、危机事件、知识搜索、知识详情与知识问答。

    以下脚手架或后续 Sprint 占位能力不纳入本次正式契约：
    tests、voice、完整 RAG。
servers:
  - url: http://localhost:8000
    description: 本地开发环境
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: 注册
      operationId: register
      description: |
        Sprint 1 注册接口。
        后端依据 `age_range` 推导 `user_mode`：
        `13_15`、`16_17` -> `teen`，其余 -> `adult`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: demo@example.com
              password: "12345678"
              age_range: 16_17
      responses:
        "200":
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              example:
                user_id: 8b6f8d6f-8b4a-4cf0-b7b1-7d41ea4d2a19
                access_token: replace-with-real-jwt
                refresh_token: replace-with-real-refresh-token
                user_mode: teen
                onboarding_completed: false
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: demo@example.com
              password: "12345678"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                access_token: replace-with-real-jwt
                refresh_token: replace-with-real-refresh-token
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: 获取当前用户
      operationId: getCurrentUser
      description: |
        Sprint 1 联调期应携带 Bearer Token。
        当前脚手架暂未校验 Authorization 头，但前后端联调仍按正式鉴权方式传递。
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 当前用户基础信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
              example:
                user_id: demo-user
                nickname: demo
                user_mode: adult
                memory_mode: summary_only
                companion_style: gentle
                voice_enabled: false
                save_voice_audio: false
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ValidationError:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
    AgeRange:
      type: string
      enum:
        - 13_15
        - 16_17
        - 18_plus
      description: 注册阶段用于推导 teen/adult 模式
    UserMode:
      type: string
      enum:
        - teen
        - adult
...
    MemoryMode:
      type: string
      enum:
        - off
        - summary_only
        - long_term
      description: 关闭、只记摘要、长期记忆。
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        age_range:
          $ref: "#/components/schemas/AgeRange"
      required:
        - email
        - password
        - age_range
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        user_mode:
          $ref: "#/components/schemas/UserMode"
        onboarding_completed:
          type: boolean
      required:
        - user_id
        - access_token
        - refresh_token
        - user_mode
        - onboarding_completed
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
      required:
        - access_token
        - refresh_token
    CurrentUserResponse:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        nickname:
          type: string
        age_range:
          $ref: "#/components/schemas/AgeRange"
        user_mode:
          $ref: "#/components/schemas/UserMode"
        usage_goals:
          type: array
          items:
            type: string
        onboarding_completed:
          type: boolean
        memory_mode:
          $ref: "#/components/schemas/MemoryMode"
        companion_style:
          type: string
        voice_enabled:
          type: boolean
        save_voice_audio:
          type: boolean
      required:
        - user_id
        - username
        - nickname
        - age_range
        - user_mode
        - usage_goals
        - onboarding_completed
        - memory_mode
        - companion_style
        - voice_enabled
        - save_voice_audio
...
    ValidationError:
      type: object
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
      required:
        - loc
        - msg
        - type
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
      required:
        - detail